Breaking apart atomic-ref tests to reduce compilation time.
